+++ spoon-maven-plugin/src/main/java/com/squareup/spoon/mojo/RunMojo.java
@@ -1,11 +1,11 @@
 // Copyright 2012 Square, Inc.
-package com.squareup.spoon;
+package com.squareup.spoon.mojo;
 
 import com.google.common.base.Strings;
+import com.squareup.spoon.SpoonRunner;
 import java.io.File;
 import java.util.List;
 import org.apache.maven.artifact.Artifact;
-import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.logging.Log;
 import org.apache.maven.plugins.annotations.Component;
@@ -37,7 +37,7 @@ import static org.apache.maven.plugins.annotations.LifecyclePhase.INTEGRATION_TE
  */
 @SuppressWarnings("UnusedDeclaration") // Used reflectively by Maven.
 @Mojo(name = "run", defaultPhase = INTEGRATION_TEST, threadSafe = false)
-public class SpoonMojo extends AbstractMojo {
+public class RunMojo extends AbstractSpoonMojo {
   private static final String SPOON_GROUP_ID = "com.squareup.spoon";
   private static final String SPOON_PLUGIN_ARTIFACT_ID = "spoon-maven-plugin";
   private static final String SPOON_RUNNER_ARTIFACT_ID = "spoon-runner";
@@ -55,10 +55,6 @@ public class SpoonMojo extends AbstractMojo {
   /** Configuration option to skip execution. */
   @Parameter
   private boolean skip;
-
-  /** Location of the output directory. */
-  @Parameter(defaultValue = "${project.build.directory}/spoon-output")
-  private File outputDirectory;
 
   /** A title for the output website. */
   @Parameter(defaultValue = "${project.name}")
@@ -106,7 +102,7 @@ public class SpoonMojo extends AbstractMojo {
   @Parameter(defaultValue = "${project.remoteProjectRepositories}", readonly = true)
   private List<RemoteRepository> remoteRepositories;
 
-  public void execute() throws MojoExecutionException {
+  @Override public void execute() throws MojoExecutionException {
     Log log = getLog();
 
     if (mavenTestSkip || mavenSkipTests || skip) {
@@ -117,10 +113,14 @@ public class SpoonMojo extends AbstractMojo {
       return;
     }
 
+    if (androidSdk == null) {
+      throw new MojoExecutionException(
+          "Could not find Android SDK. Ensure ANDROID_HOME environment variable is set.");
+    }
     File sdkFile = new File(androidSdk);
     if (!sdkFile.exists()) {
       throw new MojoExecutionException(
-          "Could not find Android SDK. Ensure ANDROID_HOME environment variable is set.");
+          String.format("Could not find Android SDK at: %s", androidSdk));
     }
     log.debug("Android SDK: " + sdkFile.getAbsolutePath());
 
@@ -175,17 +175,22 @@ public class SpoonMojo extends AbstractMojo {
     if (!"apk".equals(instrumentationArtifact.getType())) {
       throw new MojoExecutionException("Spoon can only be invoked on a module with type 'apk'.");
     }
-    return instrumentationArtifact.getFile();
+    return aetherArtifact(instrumentationArtifact).getFile();
   }
 
   private File getApplicationApk() throws MojoExecutionException {
     for (Artifact dependency : project.getDependencyArtifacts()) {
       if ("apk".equals(dependency.getType())) {
-        return dependency.getFile();
+        return aetherArtifact(dependency).getFile();
       }
     }
     throw new MojoExecutionException(
         "Could not find application. Ensure 'apk' dependency on it exists.");
+  }
+
+  private org.sonatype.aether.artifact.Artifact aetherArtifact(Artifact dep)
+      throws MojoExecutionException {
+    return resolveArtifact(dep.getGroupId(), dep.getArtifactId(), "apk", dep.getVersion());
   }
 
   private String getSpoonClasspath() throws MojoExecutionException {
@@ -195,7 +200,7 @@ public class SpoonMojo extends AbstractMojo {
     log.debug("[getSpoonClasspath] Plugin version: " + spoonVersion);
 
     org.sonatype.aether.artifact.Artifact spoonRunner =
-        resolveArtifact(SPOON_GROUP_ID, SPOON_RUNNER_ARTIFACT_ID, spoonVersion);
+        resolveArtifact(SPOON_GROUP_ID, SPOON_RUNNER_ARTIFACT_ID, "jar", spoonVersion);
     log.debug("[getSpoonClasspath] Runner artifact: " + spoonRunner);
 
     return createClasspath(spoonRunner);
@@ -212,9 +217,9 @@ public class SpoonMojo extends AbstractMojo {
   }
 
   private org.sonatype.aether.artifact.Artifact resolveArtifact(String groupId, String artifactId,
-      String version) throws MojoExecutionException {
+      String extension, String version) throws MojoExecutionException {
     ArtifactRequest request = new ArtifactRequest();
-    request.setArtifact(new DefaultArtifact(groupId, artifactId, "jar", version));
+    request.setArtifact(new DefaultArtifact(groupId, artifactId, extension, version));
     request.setRepositories(remoteRepositories);
 
     try {
@@ -253,8 +258,9 @@ public class SpoonMojo extends AbstractMojo {
         org.sonatype.aether.artifact.Artifact resolvedArtifact;
         try {
           org.sonatype.aether.artifact.Artifact nodeArtifact = node.getDependency().getArtifact();
-          resolvedArtifact = resolveArtifact(nodeArtifact.getGroupId(),
-              nodeArtifact.getArtifactId(), nodeArtifact.getVersion());
+          resolvedArtifact =
+              resolveArtifact(nodeArtifact.getGroupId(), nodeArtifact.getArtifactId(), "jar",
+                  nodeArtifact.getVersion());
         } catch (MojoExecutionException e) {
           throw new RuntimeException(e);
         }