+++ spoon-runner/src/main/java/com/squareup/spoon/html/HtmlIndex.java
@@ -1,5 +1,6 @@
 package com.squareup.spoon.html;
 
+import com.android.ddmlib.testrunner.IRemoteAndroidTestRunner;
 import com.squareup.spoon.DeviceDetails;
 import com.squareup.spoon.DeviceResult;
 import com.squareup.spoon.DeviceTest;
@@ -38,8 +39,10 @@ final class HtmlIndex {
     int totalFailure = testsRun - totalSuccess;
 
     int deviceCount = summary.getResults().size();
+    IRemoteAndroidTestRunner.TestSize testSize = summary.getTestSize();
     String started = HtmlUtils.dateToString(summary.getStarted());
-    String totalTestsRun = testsRun + " test" + (testsRun != 1 ? "s" : "");
+    String totalTestsRun = testsRun + (testSize != null ? " " + testSize.name().toLowerCase() : "")
+        + " test" + (testsRun != 1 ? "s" : "");
     String totalDevices = deviceCount + " device" + (deviceCount != 1 ? "s" : "");
 
     StringBuilder subtitle = new StringBuilder();
@@ -94,7 +97,20 @@ final class HtmlIndex {
     }
 
     @Override public int compareTo(Device other) {
+      if (name == null && other.name == null) {
+        return serial.compareTo(other.serial);
+      }
+      if (name == null) {
+        return 1;
+      }
+      if (other.name == null) {
+        return -1;
+      }
       return name.compareTo(other.name);
+    }
+
+    @Override public String toString() {
+      return name != null ? name : serial;
     }
   }